requirements.txt
pytest
pytest-html
pytest-xdist
selenium
requests
pyyaml
flask
flask-cors